Ready for core 2.6.3
