+++ networkStuff.h
@@ -3,7 +3,7 @@
 **  Program  : networkStuff.h, part of DSMRloggerWS
 **  Version  : v1.0.4
 **
-**  Copyright (c) 2019 Willem Aandewiel
+**  Copyright (c) 2020 Willem Aandewiel
 **
 **  TERMS OF USE: MIT License. See bottom of file.                                                            
 ***************************************************************************      
@@ -15,7 +15,8 @@
 #include <WiFiUdp.h>            // part of ESP8266 Core https://github.com/esp8266/Arduino
 #include <ESP8266mDNS.h>        // part of ESP8266 Core https://github.com/esp8266/Arduino
 #ifdef USE_UPDATE_SERVER
-  #include <ESP8266HTTPUpdateServer.h>
+  #include <ModUpdateServer.h>  // https://github.com/mrWheel/ModUpdateServer
+  #include "UpdateServerHtml.h"
 #endif
 #include <WiFiManager.h>        // version 0.14.0 - https://github.com/tzapu/WiFiManager
 // included in main program: #include <TelnetStream.h>       // Version 0.0.1 - https://github.com/jandrassy/TelnetStream
@@ -30,7 +31,6 @@ ESP8266WebServer        httpServer (80);
 
 WebSocketsServer webSocket = WebSocketsServer(81);
 
-bool        OtaInProgress = false;
 static      FSInfo SPIFFSinfo;
 bool        SPIFFSmounted; 
 bool        isConnected = false;
